Update the code in accordance to the checkstyle
