+++ tests/test-war-integration/src/main/java/com/google/cloud/runtimes/tomcat/test/integration/CustomServlet.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.runtimes.tomcat.test.simple;
+package com.google.cloud.runtimes.tomcat.test.integration;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -23,13 +23,24 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(urlPatterns = {"/hello/*"})
-public class HelloServlet extends HttpServlet {
+/**
+ * Reference the custom tests for the integration framework
+ * (https://github.com/GoogleCloudPlatform/runtimes-common/tree/master/integration_tests)
+ */
+@WebServlet(urlPatterns = {"/custom"})
+public class CustomServlet extends HttpServlet {
 
   @Override
-  protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-    resp.setContentType("text.plain");
-    resp.getWriter().println("Hello from the simple war application");
-  }
+  protected void doGet(HttpServletRequest req, HttpServletResponse resp)
+      throws ServletException, IOException {
+    StringBuilder configuration = new StringBuilder();
+    configuration.append("[");
+    configuration.append("{\"path\": \"custom/tests/secure\"},");
+    // Dump the server configuration into the test logs
+    configuration.append("{\"path\": \"dump/all\"}");
+    configuration.append("]");
 
+    resp.setContentType("application/json");
+    resp.getWriter().print(configuration.toString());
+  }
 }
